+++ tests/functional/Session.php
@@ -1221,4 +1221,12 @@ class Session extends \DbTestCase
 
         $this->boolean(\Session::shouldReloadActiveEntities())->isFalse();
     }
+
+    public function testActiveEntityNameForFullStructure(): void
+    {
+        $this->login();
+        \Session::changeActiveEntities("all");
+        $this->string($_SESSION["glpiactive_entity_name"])->isEqualTo("Root entity (full structure)");
+        $this->string($_SESSION["glpiactive_entity_shortname"])->isEqualTo("Root entity (full structure)");
+    }
 }